requirements.txt
# Production dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiogram>=3.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
gspread>=5.0.0
google-auth>=2.0.0
apscheduler>=3.10.0
structlog>=23.0.0
python-multipart>=0.0.6
pytz>=2023.3